+++ typescript/metrics/src/balance.ts
@@ -153,13 +153,18 @@ export async function getXERC20Info(
   warpCore: WarpCore,
   token: Token,
 ): Promise<XERC20Info> {
-  if (token.protocol !== ProtocolType.Ethereum) {
+  if (
+    token.protocol !== ProtocolType.Ethereum &&
+    token.protocol !== ProtocolType.Tron
+  ) {
     throw new Error(`Unsupported XERC20 protocol type ${token.protocol}`);
   }
 
   if (
     token.standard === TokenStandard.EvmHypXERC20 ||
-    token.standard === TokenStandard.EvmHypVSXERC20
+    token.standard === TokenStandard.EvmHypVSXERC20 ||
+    token.standard == TokenStandard.TronHypVSXERC20 ||
+    token.standard == TokenStandard.TronHypXERC20
   ) {
     const adapter = token.getAdapter(
       warpCore.multiProvider,
@@ -170,7 +175,9 @@ export async function getXERC20Info(
     };
   } else if (
     token.standard === TokenStandard.EvmHypXERC20Lockbox ||
-    token.standard === TokenStandard.EvmHypVSXERC20Lockbox
+    token.standard === TokenStandard.EvmHypVSXERC20Lockbox ||
+    token.standard === TokenStandard.TronHypXERC20Lockbox ||
+    token.standard === TokenStandard.TronHypVSXERC20Lockbox
   ) {
     const adapter = token.getAdapter(
       warpCore.multiProvider,