+++ typescript/provider-sdk/src/protocol.ts
@@ -21,6 +21,7 @@ export enum ProtocolType {
   Starknet = 'starknet',
   Radix = 'radix',
   Aleo = 'aleo',
+  Tron = 'tron',
   Unknown = 'unknown',
 }
 
@@ -35,6 +36,7 @@ export const ProtocolSmallestUnit = {
   [ProtocolType.Starknet]: 'fri',
   [ProtocolType.Radix]: 'attos',
   [ProtocolType.Aleo]: 'microcredits',
+  [ProtocolType.Tron]: 'SUN',
   [ProtocolType.Unknown]: 'unknown',
 };