+++ typescript/sdk/src/deploy/warp.ts
@@ -212,6 +212,7 @@ export async function executeWarpDeploy(
     }
 
     switch (protocol) {
+      case ProtocolType.Tron:
       case ProtocolType.Ethereum: {
         const deployer = warpDeployConfig.isNft
           ? new HypERC721Deployer(multiProvider)
@@ -356,6 +357,7 @@ async function createWarpIsm({
   const protocolType = multiProvider.getProtocol(chain);
 
   switch (protocolType) {
+    case ProtocolType.Tron:
     case ProtocolType.Ethereum: {
       const evmIsmModule = await EvmIsmModule.create({
         chain,
@@ -420,6 +422,7 @@ async function createWarpHook({
   const protocolType = multiProvider.getProtocol(chain);
 
   switch (protocolType) {
+    case ProtocolType.Tron:
     case ProtocolType.Ethereum: {
       rootLogger.info(`Loading registry factory addresses for ${chain}...`);
 
@@ -534,6 +537,7 @@ export async function enrollCrossChainRouters(
       let transactions: TypedAnnotatedTransaction[] = [];
 
       switch (protocol) {
+        case ProtocolType.Tron:
         case ProtocolType.Ethereum: {
           const {
             domainRoutingIsmFactory,