+++ typescript/sdk/src/token/tokenMetadataUtils.ts
@@ -4,7 +4,7 @@ import {
   IERC4626__factory,
   IXERC20Lockbox__factory,
 } from '@hyperlane-xyz/core';
-import { ProtocolType } from '@hyperlane-xyz/utils';
+import { isEVMLike } from '@hyperlane-xyz/utils';
 
 import { MultiProvider } from '../providers/MultiProvider.js';
 
@@ -42,7 +42,7 @@ export async function deriveTokenMetadata(
       metadataMap.set(chain, TokenMetadataSchema.parse(config));
     }
 
-    if (multiProvider.getProtocol(chain) !== ProtocolType.Ethereum) {
+    if (!isEVMLike(multiProvider.getProtocol(chain))) {
       continue;
     }