+++ typescript/cosmos-types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/cosmos-types
 
+## 26.0.0
+
 ## 25.5.0
 
 ## 25.4.1