+++ typescript/metrics/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @hyperlane-xyz/metrics
 
+## 0.2.0
+
+### Minor Changes
+
+- 1d116d8: Added Tron ProtocolType & deprecated Tron TechnicalStack. Add support for TronLink wallet in the widgets.
+
+### Patch Changes
+
+- Updated dependencies [43255a9]
+- Updated dependencies [06aacac]
+- Updated dependencies [763a264]
+- Updated dependencies [1d116d8]
+  - @hyperlane-xyz/sdk@26.0.0
+  - @hyperlane-xyz/utils@26.0.0
+  - @hyperlane-xyz/core@11.0.1
+
 ## 0.1.11
 
 ### Patch Changes