+++ typescript/radix-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hyperlane-xyz/radix-sdk
 
+## 26.0.0
+
+### Patch Changes
+
+- Updated dependencies [06aacac]
+- Updated dependencies [1d116d8]
+  - @hyperlane-xyz/utils@26.0.0
+  - @hyperlane-xyz/provider-sdk@3.0.0
+
 ## 25.5.0
 
 ### Patch Changes